requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.2.7            # via django
certifi==2020.4.5.1       # via requests
cffi==1.14.0              # via cryptography
chardet==3.0.4            # via requests
click==7.1.2              # via pip-tools
coreapi==2.3.3            # via django-rest-swagger, openapi-codec
coreschema==0.0.4         # via coreapi
cryptography==2.9.2       # via social-auth-core
defusedxml==0.6.0         # via python3-openid, python3-saml, social-auth-core
django-crispy-forms==1.9.0  # via -r requirements.in
django-rest-swagger==2.2.0  # via -r requirements.in
django==3.0.6             # via -r requirements.in, djangorestframework
djangorestframework==3.11.0  # via -r requirements.in, django-rest-swagger
idna==2.9                 # via requests
isodate==0.6.0            # via python3-saml
itypes==1.2.0             # via coreapi
jinja2==2.11.2            # via coreschema
lxml==4.5.0               # via xmlsec
markupsafe==1.1.1         # via jinja2
mysqlclient==1.4.6        # via -r requirements.in
oauthlib==3.1.0           # via requests-oauthlib, social-auth-core
openapi-codec==1.3.2      # via django-rest-swagger
pip-tools==5.1.2          # via -r requirements.in
pkgconfig==1.5.1          # via xmlsec
pycparser==2.20           # via cffi
pyjwt==1.7.1              # via social-auth-core
python-memcached==1.59    # via -r requirements.in
python3-openid==3.1.0     # via social-auth-core
python3-saml==1.9.0       # via -r requirements.in, social-auth-core
pytz==2020.1              # via django
pyyaml==5.3.1             # via -r requirements.in
requests-oauthlib==1.3.0  # via social-auth-core
requests==2.23.0          # via coreapi, requests-oauthlib, social-auth-core
simplejson==3.17.0        # via django-rest-swagger
six==1.14.0               # via cryptography, isodate, pip-tools, python-memcached, social-auth-app-django, social-auth-core
social-auth-app-django==3.1.0  # via -r requirements.in
social-auth-core[saml]==3.3.3  # via -r requirements.in, social-auth-app-django
sqlparse==0.3.1           # via django
uritemplate==3.0.1        # via coreapi
urllib3==1.25.9           # via requests
xmlsec==1.3.3             # via python3-saml

# The following packages are considered to be unsafe in a requirements file:
# pip
